requirements.txt
fastapi
uvicorn
llama-index
azure-search-documents
streamlit
PyMuPDF
msal
pytest
ruff
python-dotenv
azure-identity
requests
tiktoken
openai
python-dateutil
